fix: padronizando os regex de login e cadastro
