--- src/components/Register/functions/checkPassword.ts
@@ -1,8 +1,0 @@
-export const checkPassWord = (password: string): boolean => {
-  const regex = new RegExp(
-    `^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%^&*()_+\\-=[\\]{};':"|,.<>/?]).{8,16}$`,
-    'g'
-  );
-
-  return regex.test(password);
-};
